requirements.txt
fastapi
uvicorn[standard]
python-multipart
pyyaml
python-dotenv
pandas
chromadb
langchain
langchain-chroma
langchain-huggingface
langchain-openai
langchain-text-splitters
sentence-transformers